backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.6

# Google AI (Gemini)
google-generativeai==0.3.2

# Environment Variables
python-dotenv==1.0.0

# Redis (for caching and sessions)
redis==5.0.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# HTTP Requests
httpx==0.26.0
requests==2.31.0

# Date & Time
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Monitoring
prometheus-client==0.19.0

# CORS
python-cors==1.0.0